perf_tests: remove false dependence on google test
