+++ batched/dense/unit_test/Test_Batched_SerialSVD.hpp
@@ -18,6 +18,8 @@
 #include "KokkosBatched_SVD_Decl.hpp"             //For testing overall kernel
 #include "KokkosBatched_SVD_Serial_Internal.hpp"  //For unit testing individual components
 #include "KokkosBatched_SetIdentity_Decl.hpp"
+#include "KokkosKernels_TestMatrixUtils.hpp"
+#include "KokkosKernels_TestVanilla.hpp"
 
 namespace Test {
 template <typename Scalar>
