+++ perf_test/CMakeLists.txt
@@ -11,21 +11,6 @@ if (KokkosKernels_ENABLE_PERFTESTS)
     #build correctly with or without MPI, but only run them with a single
     #MPI process.
 
-    SET(GTEST_SOURCE_DIR ${PACKAGE_SOURCE_DIR}/tpls/gtest)
-
-    KOKKOSKERNELS_ADD_TEST_LIBRARY(
-                                kokkoskernelsperf_gtest
-                                HEADERS ${GTEST_SOURCE_DIR}/gtest/gtest.h
-                                SOURCES ${GTEST_SOURCE_DIR}/gtest/gtest-all.cc
-                                    )
-    #Disables pthreads, this is a problem for serial builds in Trilinos & Sierra if it's enabled.
-
-    TARGET_COMPILE_DEFINITIONS(kokkoskernelsperf_gtest PUBLIC "-DGTEST_HAS_PTHREAD=0")
-    TARGET_INCLUDE_DIRECTORIES(kokkoskernelsperf_gtest PUBLIC $<BUILD_INTERFACE:${GTEST_SOURCE_DIR}>)
-
-    #Gtest minimally requires C++ 11
-    TARGET_COMPILE_FEATURES(kokkoskernelsperf_gtest PUBLIC cxx_std_11)
-
     KOKKOSKERNELS_INCLUDE_DIRECTORIES(sparse)
 
     if(KokkosKernels_ENABLE_TESTS_AND_PERFSUITE)
