+++ perf_test/KokkosKernels_perf_test_utilities.hpp
@@ -20,7 +20,7 @@
 #ifndef KOKKOSKERNELS_PERF_TEST_UTILITIES_HPP
 #define KOKKOSKERNELS_PERF_TEST_UTILITIES_HPP
 
-#include "KokkosKernels_TestUtils.hpp"  // for string_compare_no_case
+#include "KokkosKernels_TestStringUtils.hpp"  // for string_compare_no_case
 
 // Namepsace that defines common utilities
 // for performance tests
