+++ perf_test/blas/blas1/KokkosBlas_dot_mv_perf_test.cpp
@@ -14,10 +14,12 @@
 //
 //@HEADER
 
+#include <iostream>
+
 #include <Kokkos_Core.hpp>
 #include <src/KokkosBlas1_dot.hpp>
 #include <Kokkos_Random.hpp>
-#include "KokkosKernels_TestUtils.hpp"
+#include "KokkosKernels_TestStringUtils.hpp"
 
 struct Params {
   int use_cuda    = 0;
