+++ perf_test/blas/blas2/KokkosBlas2_ger_perf_test_benchmark.cpp
@@ -21,7 +21,7 @@
 #include "KokkosBlas2_ger.hpp"
 #include <typeinfo>
 
-#include "KokkosKernels_TestUtils.hpp"
+#include "KokkosKernels_TestStringUtils.hpp"
 #include "KokkosKernels_perf_test_utilities.hpp"
 
 #include <Benchmark_Context.hpp>
