+++ perf_test/blas/blas3/CMakeLists.txt
@@ -4,7 +4,6 @@ KOKKOSKERNELS_INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR})
 KOKKOSKERNELS_ADD_EXECUTABLE(
     KokkosBlas3_perf_test
     SOURCES KokkosBlas3_perf_test.cpp
-    TESTONLYLIBS kokkoskernelsperf_gtest
 )
 
 KOKKOSKERNELS_ADD_EXECUTABLE(
