+++ perf_test/blas/blas3/KokkosBlas3_gemm_perf_test.hpp
@@ -28,8 +28,10 @@
 #include "KokkosBatched_HostLevel_Gemm.hpp"
 #include "KokkosBatched_Gemm_Decl.hpp"
 #include "KokkosBatched_Util.hpp"
-#include "gtest/gtest.h"  // EXPECT_NEAR
-#include "KokkosKernels_TestUtils.hpp"
+// #include "gtest/gtest.h"  // EXPECT_NEAR
+#include "KokkosKernels_TestStringUtils.hpp"
+#include "KokkosKernels_TestEpsilon.hpp"
+#include "KokkosKernels_TestVanilla.hpp"
 
 #include <chrono>
